app/requirements.txt
django
numpy
djangorestframework
django-cors-headers
python-decouple
gunicorn
whitenoise
psycopg2-binary